Copy xhr logic to a different file
